LMS: Configurator: drop only new tasks during stopping
TaskCompleted removes global flags 'm_inProcess' and run the next task.
If m_inProcess is not set, it safe to remove all new incoming tasks
by calling TaskCompleted() with conjunction of m_needToStop set.
On the other hand, if the flag is set it, calling TaskCompleted()
would kill currently running task, which we wish to finish
as it is involved in stopping.

Add a print of Operation that is being refreshed

Signed-off-by: Riki Sharon <riki.sharon@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>