LMS: Configurator: avoid toggling services that does not exists
In the case of request to start/stop non-existent service
the execution chain will stuck.
This can happen if asking to toggle state of the service that
do not exists on the current OS.
Compile out such toggles under ifdef guard for right OS
and mark such events as processed on OS where service not exists.

Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>