LMS: PortForwadingService: call DeInit in LMEConnection destructor
Currently LMEConnection relies on a parent class call of DeInit
before destruction.
If the parent class fail to call DeInit for some reason
(like timing issue in Debug mode) the thread is left running
and process crashes on access to destructed object.
Call DeInit explicitly in destructor to close that gap.

Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>