LMS: GMS_COMMON: fix certificate bypass
CryptQueryObject always returns just the first certificate even in case
of dual signing, regardless of the signature verification status. Since
the DLL was built in a way that the Intel certificate (without
signature) is first - this is returned, and verify as Intel owned

WinVerifyTrust, unless specifically told to, will check ALL signatures,
and as long as one of the signatures is valid it will return success.
Since the secondary signature was valid, test passes and continue to
load DLL.

Since Win8, the WINTRUST_DATA data structure has an additional field
that allows choosing which certificate to use to verify the signature
with, and will also return if there are additional signing certificates.

Since we only care about the first signature, we don't have to check the
additional signatures.

Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>